chore: realign py response types
